rename: fixutures/board => fixutures/qnaboard 변경
